requirements.txt
ShopifyAPI
boto3
simplejson
